Added using of HttpContext
